+++ src/plotter_disk.hpp
@@ -29,9 +29,6 @@
 #include <iostream>
 #include <map>
 #include <string>
-#include <unordered_map>
-#include <unordered_set>
-#include <utility>
 #include <vector>
 
 // Gulrak filesystem brings in Windows headers that cause some issues with std
@@ -43,555 +40,13 @@ namespace fs = ghc::filesystem;
 #include "calculate_bucket.hpp"
 #include "encoding.hpp"
 #include "pos_constants.hpp"
-#include "fast_sort_on_disk.hpp"
+#include "sort_manager.hpp"
 #include "util.hpp"
-#include "threading.hpp"
+#include "phase1.hpp"
+#include "phase2.hpp"
+#include "phase3.hpp"
+#include "phase4.hpp"
 
-// Constants that are only relevant for the plotting process.
-// Other constants can be found in pos_constants.hpp
-
-// Distance between matching entries is stored in the offset
-const uint32_t kOffsetSize = 10;
-
-// During backprop and compress, the write pointer is ahead of the read pointer
-// Note that the large the offset, the higher these values must be
-const uint32_t kReadMinusWrite = 1U << kOffsetSize;
-const uint32_t kCachedPositionsSize = kReadMinusWrite * 4;
-
-// Max matches a single entry can have, used for hardcoded memory allocation
-const uint32_t kMaxMatchesSingleEntry = 30;
-const uint32_t kMinBuckets = 16;
-const uint32_t kMaxBuckets = 128;
-
-// TODO: remove
-const uint32_t kNumSortBuckets = 16;
-const uint32_t kLogNumSortBuckets = 4;
-
-const double kMemSortProportion = 0.75;
-
-// Results of phase 3. These are passed into Phase 4, so the checkpoint tables
-// can be properly built.
-struct Phase3Results {
-    // Pointers to each table start byet in the final file
-    vector<uint64_t> final_table_begin_pointers;
-    // Number of entries written for f7
-    uint64_t final_entries_written;
-    uint32_t right_entry_size_bits;
-
-    uint32_t header_size;
-    LazySortManager* table7_sm;
-};
-
-typedef struct {
-    int index;
-#ifdef _WIN32
-    HANDLE mine;
-    HANDLE theirs;
-#else
-    sem_t* mine;
-    sem_t* theirs;
-#endif
-    uint64_t right_entry_size_bytes;
-    uint8_t k;
-    uint8_t table_index;
-    uint8_t metadata_size;
-    uint32_t entry_size_bytes;
-    uint8_t pos_size;
-    uint64_t prevtableentries;
-    uint32_t compressed_entry_size_bytes;
-    std::vector<FileDisk>* ptmp_1_disks;
-} THREADDATA;
-
-struct GlobalData {
-    uint64_t left_writer_count;
-    uint64_t right_writer_count;
-    uint64_t matches;
-    LazySortManager* L_sort_manager;
-    LazySortManager* R_sort_manager;
-    uint64_t left_reader_buf_size;
-    uint64_t right_writer_buf_size;
-    uint64_t left_writer_buf_size;
-    uint8_t *left_reader_buf;
-    uint8_t *left_writer_buf;
-    uint64_t left_writer_buf_entries;
-    uint64_t right_writer_buf_entries;
-    uint64_t left_writer;
-    uint64_t right_writer;
-    uint64_t stripe_size;
-    uint8_t num_threads;
-};
-
-GlobalData globals;
-
-
-PlotEntry GetLeftEntry(
-    uint8_t table_index,
-    uint8_t* left_buf,
-    uint8_t k,
-    uint8_t metadata_size,
-    uint8_t pos_size)
-{
-    PlotEntry left_entry;
-    left_entry.y = 0;
-    left_entry.read_posoffset = 0;
-    left_entry.left_metadata = 0;
-    left_entry.right_metadata = 0;
-
-    uint32_t ysize = (table_index == 7) ? k : k + kExtraBits;
-
-    if (table_index == 1) {
-        // For table 1, we only have y and metadata
-        left_entry.y = Util::SliceInt64FromBytes(left_buf, 0, k + kExtraBits);
-        left_entry.left_metadata =
-            Util::SliceInt64FromBytes(left_buf, k + kExtraBits, metadata_size);
-    } else {
-        // For tables 2-6, we we also have pos and offset. We need to read this because
-        // this entry will be written again to the table without the y (and some entries
-        // are dropped).
-        left_entry.y = Util::SliceInt64FromBytes(left_buf, 0, ysize);
-        left_entry.read_posoffset =
-            Util::SliceInt64FromBytes(left_buf, ysize, pos_size + kOffsetSize);
-        if (metadata_size <= 128) {
-            left_entry.left_metadata =
-                Util::SliceInt128FromBytes(left_buf, ysize + pos_size + kOffsetSize, metadata_size);
-        } else {
-            // Large metadatas that don't fit into 128 bits. (k > 32).
-            left_entry.left_metadata =
-                Util::SliceInt128FromBytes(left_buf, ysize + pos_size + kOffsetSize, 128);
-            left_entry.right_metadata = Util::SliceInt128FromBytes(
-                left_buf, ysize + pos_size + kOffsetSize + 128, metadata_size - 128);
-        }
-    }
-    return left_entry;
-}
-
-#ifdef _WIN32
-DWORD WINAPI thread(LPVOID lpParameter)
-{
-    THREADDATA* ptd = (THREADDATA*)lpParameter;
-#else
-void* thread(void* arg)
-{
-    THREADDATA* ptd = (THREADDATA*)arg;
-#endif
-
-    std::vector<uint64_t> right_bucket_sizes(kNumSortBuckets, 0);
-
-    uint64_t right_entry_size_bytes = ptd->right_entry_size_bytes;
-    uint8_t k = ptd->k;
-    uint8_t table_index = ptd->table_index;
-    uint8_t metadata_size = ptd->metadata_size;
-    uint32_t entry_size_bytes = ptd->entry_size_bytes;
-    uint8_t pos_size = ptd->pos_size;
-    uint64_t prevtableentries = ptd->prevtableentries;
-    uint32_t compressed_entry_size_bytes = ptd->compressed_entry_size_bytes;
-    std::vector<FileDisk>* ptmp_1_disks = ptd->ptmp_1_disks;
-
-    // Streams to read and right to tables. We will have handles to two tables. We will
-    // read through the left table, compute matches, and evaluate f for matching entries,
-    // writing results to the right table.
-    uint64_t left_buf_entries = (uint64_t)(globals.stripe_size) + 2500;
-    uint64_t right_buf_entries = (uint64_t)(globals.stripe_size) + 2500;
-    uint8_t* right_writer_buf = new uint8_t[right_buf_entries * right_entry_size_bytes];
-    uint8_t* left_writer_buf = new uint8_t[left_buf_entries * compressed_entry_size_bytes];
-
-    FxCalculator f(k, table_index + 1);
-
-    // Stores map of old positions to new positions (positions after dropping entries from L
-    // table that did not match) Map ke
-    uint16_t position_map_size = 2000;
-
-    uint16_t* L_position_map =
-        new uint16_t[position_map_size];  // Should comfortably fit 2 buckets worth of items
-    uint16_t* R_position_map = new uint16_t[position_map_size];
-
-    // Start at left table pos = 0 and iterate through the whole table. Note that the left table
-    // will already be sorted by y
-    uint64_t totalstripes = (prevtableentries + globals.stripe_size - 1) / globals.stripe_size;
-    uint64_t threadstripes = (totalstripes + globals.num_threads - 1) / globals.num_threads;
-
-    std::vector<Bits> to_write_R_entries;
-
-    for (uint64_t stripe = 0; stripe < threadstripes; stripe++) {
-        uint64_t pos = (stripe * globals.num_threads + ptd->index) * globals.stripe_size;
-        uint64_t endpos = pos + globals.stripe_size + 1;  // one y value overlap
-        uint64_t left_reader = pos * entry_size_bytes;
-        uint64_t left_reader_prev_stripe = (pos - globals.stripe_size) * entry_size_bytes;
-        uint64_t left_writer_count = 0;
-        uint64_t stripe_left_writer_count = 0;
-        uint64_t stripe_start_correction = 0xffffffffffffffff;
-        uint64_t right_writer_count = 0;
-        uint64_t matches = 0;  // Total matches
-
-        // This is a sliding window of entries, since things in bucket i can match with things in
-        // bucket
-        // i + 1. At the end of each bucket, we find matches between the two previous buckets.
-        std::vector<PlotEntry> bucket_L;
-        std::vector<PlotEntry> bucket_R;
-
-        uint64_t bucket = 0;
-        bool end_of_table = false;  // We finished all entries in the left table
-
-        uint64_t ignorebucket = 0xffffffffffffffff;
-        bool bMatch = false;
-        bool bFirstStripeOvertimePair = false;
-        bool bSecondStripOvertimePair = false;
-        bool bThirdStripeOvertimePair = false;
-
-        bool bStripePregamePair = false;
-        bool bStripeStartPair = false;
-
-        uint64_t L_position_base = 0;
-        uint64_t R_position_base = 0;
-        uint64_t newlpos = 0;
-        uint64_t newrpos = 0;
-        Bits new_left_entry(0, pos_size + kOffsetSize);
-        std::vector<std::tuple<PlotEntry, PlotEntry, std::pair<Bits, Bits>>>
-            current_entries_to_write;
-        std::vector<std::tuple<PlotEntry, PlotEntry, std::pair<Bits, Bits>>>
-            future_entries_to_write;
-        std::vector<std::pair<uint16_t, uint16_t>> match_indexes;
-        std::vector<PlotEntry*> not_dropped;  // Pointers are stored to avoid copying entries
-
-        if (pos == 0) {
-            bMatch = true;
-            bStripePregamePair = true;
-            bStripeStartPair = true;
-            stripe_left_writer_count = 0;
-            stripe_start_correction = 0;
-        }
-
-
-        bool need_new_bucket = globals.L_sort_manager->CloseToNewBucket(left_reader);
-
-        if (need_new_bucket) {
-            bool will_trigger_bucket = !globals.L_sort_manager->CloseToNewBucket(left_reader_prev_stripe);
-            SemaphoreUtils::Wait(ptd->theirs);
-            if (will_trigger_bucket) {
-                globals.L_sort_manager->TriggerNewBucket(left_reader, 0);
-            }
-        }
-
-        while (pos < prevtableentries + 1) {
-            PlotEntry left_entry = PlotEntry();
-            if (pos >= prevtableentries) {
-                end_of_table = true;
-                left_entry.y = 0;
-                left_entry.left_metadata = 0;
-                left_entry.right_metadata = 0;
-                left_entry.used = false;
-            } else {
-                // Reads a left entry from disk
-                uint8_t* left_buf = globals.L_sort_manager->ReadEntry(left_reader);
-                left_reader += entry_size_bytes;
-
-                left_entry = GetLeftEntry(table_index, left_buf, k, metadata_size, pos_size);
-            }
-
-            // This is not the pos that was read from disk,but the position of the entry we read,
-            // within L table.
-            left_entry.pos = pos;
-            left_entry.used = false;
-            uint64_t y_bucket = left_entry.y / kBC;
-
-            if (!bMatch) {
-                if (ignorebucket == 0xffffffffffffffff) {
-                    ignorebucket = y_bucket;
-                } else {
-                    if ((y_bucket != ignorebucket)) {
-                        bucket = y_bucket;
-                        bMatch = true;
-                    }
-                }
-            }
-            if (!bMatch) {
-                stripe_left_writer_count++;
-                R_position_base = stripe_left_writer_count;
-                pos++;
-                continue;
-            }
-
-            // Keep reading left entries into bucket_L and R, until we run out of things
-            if (y_bucket == bucket) {
-                bucket_L.emplace_back(left_entry);
-            } else if (y_bucket == bucket + 1) {
-                bucket_R.emplace_back(left_entry);
-            } else {
-                // cout << "matching! " << bucket << " and " << bucket + 1 << endl;
-                // This is reached when we have finished adding stuff to bucket_R and bucket_L,
-                // so now we can compare entries in both buckets to find matches. If two entries
-                // match, match, the result is written to the right table. However the writing
-                // happens in the next iteration of the loop, since we need to remap positions.
-                if (!bucket_L.empty()) {
-                    not_dropped.clear();
-
-                    if (!bucket_R.empty()) {
-                        // Compute all matches between the two buckets and save indeces.
-                        match_indexes = f.FindMatches(bucket_L, bucket_R);
-
-                        // We mark entries as used if they took part in a match.
-                        for (auto& indeces : match_indexes) {
-                            bucket_L[std::get<0>(indeces)].used = true;
-                            if (end_of_table) {
-                                bucket_R[std::get<1>(indeces)].used = true;
-                            }
-                        }
-                    }
-
-                    // Adds L_bucket entries that are used to not_dropped. They are used if they
-                    // either matched with something to the left (in the previous iteration), or
-                    // matched with something in bucket_R (in this iteration).
-                    for (size_t bucket_index = 0; bucket_index < bucket_L.size(); bucket_index++) {
-                        PlotEntry& L_entry = bucket_L[bucket_index];
-                        if (L_entry.used) {
-                            not_dropped.emplace_back(&bucket_L[bucket_index]);
-                        }
-                    }
-                    if (end_of_table) {
-                        // In the last two buckets, we will not get a chance to enter the next
-                        // iteration due to breaking from loop. Therefore to write the final
-                        // bucket in this iteration, we have to add the R entries to the
-                        // not_dropped list.
-                        for (size_t bucket_index = 0; bucket_index < bucket_R.size();
-                             bucket_index++) {
-                            PlotEntry& R_entry = bucket_R[bucket_index];
-                            if (R_entry.used) {
-                                not_dropped.emplace_back(&R_entry);
-                            }
-                        }
-                    }
-                    // We keep maps from old positions to new positions. We only need two maps,
-                    // one for L bucket and one for R bucket, and we cycle through them. Map
-                    // keys are stored as positions % 2^10 for efficiency. Map values are stored
-                    // as offsets from the base position for that bucket, for efficiency.
-                    uint16_t* tmp = L_position_map;
-                    L_position_map = R_position_map;
-                    R_position_map = tmp;
-                    L_position_base = R_position_base;
-                    R_position_base = stripe_left_writer_count;
-
-                    for (PlotEntry*& entry : not_dropped) {
-                        // The new position for this entry = the total amount of thing written
-                        // to L so far. Since we only write entries in not_dropped, about 14% of
-                        // entries are dropped.
-                        R_position_map[entry->pos % position_map_size] =
-                            stripe_left_writer_count - R_position_base;
-
-                        if (bStripeStartPair) {
-                            if (stripe_start_correction == 0xffffffffffffffff) {
-                                stripe_start_correction = stripe_left_writer_count;
-                            }
-
-                            if(left_writer_count>=left_buf_entries) {
-                                cout << "left_writer_count overrun" << endl;
-                                exit(1);
-                            }
-                            uint8_t* tmp_buf =
-                                left_writer_buf + left_writer_count *
-                                                      compressed_entry_size_bytes;
-
-                            left_writer_count++;
-                            // memset(tmp_buf, 0xff, compressed_entry_size_bytes);
-
-                            // Rewrite left entry with just pos and offset, to reduce working space
-                            Bits new_left_entry = Bits(
-                                (table_index == 1) ? entry->left_metadata : entry->read_posoffset,
-                                (table_index == 1) ? k : pos_size + kOffsetSize);
-
-                            new_left_entry.ToBytes(tmp_buf);
-                        }
-                        stripe_left_writer_count++;
-                    }
-
-                    // Two vectors to keep track of things from previous iteration and from this
-                    // iteration.
-                    current_entries_to_write.swap(future_entries_to_write);
-                    future_entries_to_write.clear();
-
-                    for (auto& indeces : match_indexes) {
-                        PlotEntry& L_entry = bucket_L[std::get<0>(indeces)];
-                        PlotEntry& R_entry = bucket_R[std::get<1>(indeces)];
-
-                        if (bStripeStartPair)
-                            matches++;
-
-                        // Sets the R entry to used so that we don't drop in next iteration
-                        R_entry.used = true;
-                        // Computes the output pair (fx, new_metadata)
-                        if (metadata_size <= 128) {
-                            const std::pair<Bits, Bits>& f_output = f.CalculateBucket(
-                                Bits(L_entry.y, k + kExtraBits),
-                                Bits(R_entry.y, k + kExtraBits),
-                                Bits(L_entry.left_metadata, metadata_size),
-                                Bits(R_entry.left_metadata, metadata_size));
-                            future_entries_to_write.push_back(
-                                std::make_tuple(L_entry, R_entry, f_output));
-                        } else {
-                            // Metadata does not fit into 128 bits
-                            const std::pair<Bits, Bits>& f_output = f.CalculateBucket(
-                                Bits(L_entry.y, k + kExtraBits),
-                                Bits(R_entry.y, k + kExtraBits),
-                                Bits(L_entry.left_metadata, 128) +
-                                    Bits(L_entry.right_metadata, metadata_size - 128),
-                                Bits(R_entry.left_metadata, 128) +
-                                    Bits(R_entry.right_metadata, metadata_size - 128));
-                            future_entries_to_write.push_back(
-                                std::make_tuple(L_entry, R_entry, f_output));
-                        }
-                    }
-
-                    // At this point, future_entries_to_write contains the matches of buckets L
-                    // and R, and current_entries_to_write contains the matches of L and the
-                    // bucket left of L. These are the ones that we will write.
-                    uint16_t final_current_entry_size = current_entries_to_write.size();
-                    if (end_of_table) {
-                        // For the final bucket, write the future entries now as well, since we
-                        // will break from loop
-                        current_entries_to_write.insert(
-                            current_entries_to_write.end(),
-                            future_entries_to_write.begin(),
-                            future_entries_to_write.end());
-                    }
-                    for (size_t i = 0; i < current_entries_to_write.size(); i++) {
-                        const auto& entry_tuple = current_entries_to_write[i];
-                        const PlotEntry& L_entry = std::get<0>(entry_tuple);
-                        const PlotEntry& R_entry = std::get<1>(entry_tuple);
-
-                        const std::pair<Bits, Bits>& f_output = std::get<2>(entry_tuple);
-                        // We only need k instead of k + kExtraBits bits for the last table
-                        Bits new_entry = table_index + 1 == 7 ? std::get<0>(f_output).Slice(0, k)
-                                                              : std::get<0>(f_output);
-
-                        // Maps the new positions. If we hit end of pos, we must write things in
-                        // both final_entries to write and current_entries_to_write, which are
-                        // in both position maps.
-                        if (!end_of_table || i < final_current_entry_size) {
-                            newlpos =
-                                L_position_map[L_entry.pos % position_map_size] + L_position_base;
-                        } else {
-                            newlpos =
-                                R_position_map[L_entry.pos % position_map_size] + R_position_base;
-                        }
-                        newrpos = R_position_map[R_entry.pos % position_map_size] + R_position_base;
-                        // Position in the previous table
-                        new_entry.AppendValue(newlpos, pos_size);
-
-                        // Offset for matching entry
-                        if (newrpos - newlpos > (1U << kOffsetSize) * 97 / 100) {
-                            std::cout << "Offset: " << newrpos - newlpos << std::endl;
-                            abort();
-                        }
-
-                        new_entry.AppendValue(newrpos - newlpos, kOffsetSize);
-                        // New metadata which will be used to compute the next f
-                        new_entry += std::get<1>(f_output);
-
-                        if(right_writer_count>=right_buf_entries){
-                            cout << "right_writer_count overrun" << endl;
-                            exit(1);
-                        }
-
-                        if (bStripeStartPair) {
-                            uint8_t *right_buf =
-                                    right_writer_buf +
-                                    right_writer_count * right_entry_size_bytes;
-                            new_entry.ToBytes(right_buf);
-                            right_writer_count++;
-                        }
-                    }
-                }
-
-                if (pos >= endpos) {
-                    if (!bFirstStripeOvertimePair)
-                        bFirstStripeOvertimePair = true;
-                    else if (!bSecondStripOvertimePair)
-                        bSecondStripOvertimePair = true;
-                    else if (!bThirdStripeOvertimePair)
-                        bThirdStripeOvertimePair = true;
-                    else {
-                        break;
-                    }
-                } else {
-                    if (!bStripePregamePair)
-                        bStripePregamePair = true;
-                    else if (!bStripeStartPair)
-                        bStripeStartPair = true;
-                }
-
-                if (y_bucket == bucket + 2) {
-                    // We saw a bucket that is 2 more than the current, so we just set L = R, and R
-                    // = [entry]
-                    bucket_L = bucket_R;
-                    bucket_R = std::vector<PlotEntry>();
-                    bucket_R.emplace_back(std::move(left_entry));
-                    ++bucket;
-                } else {
-                    // We saw a bucket that >2 more than the current, so we just set L = [entry],
-                    // and R = []
-                    bucket = y_bucket;
-                    bucket_L = std::vector<PlotEntry>();
-                    bucket_L.emplace_back(std::move(left_entry));
-                    bucket_R = std::vector<PlotEntry>();
-                }
-            }
-            // Increase the read pointer in the left table, by one
-            ++pos;
-        }
-
-        if (!need_new_bucket) {
-            // If we needed new bucket, we already waited
-            SemaphoreUtils::Wait(ptd->theirs);
-        }
-
-        uint32_t ysize = (table_index + 1 == 7) ? k : k + kExtraBits;
-        uint32_t startbyte = ysize / 8;
-        uint32_t endbyte = (ysize + pos_size + 7) / 8 - 1;
-        uint64_t shiftamt = (8 - ((ysize + pos_size) % 8)) % 8;
-        uint64_t correction = (globals.left_writer_count - stripe_start_correction) << shiftamt;
-
-        // Correct positions
-        for (uint32_t i = 0; i < right_writer_count; i++) {
-            uint64_t posaccum = 0;
-            uint8_t* entrybuf = right_writer_buf + i * right_entry_size_bytes;
-
-            for (uint32_t j = startbyte; j <= endbyte; j++) {
-                posaccum = (posaccum << 8) | (entrybuf[j]);
-            }
-            posaccum += correction;
-            for (uint32_t j = endbyte; j >= startbyte; --j) {
-                entrybuf[j] = posaccum & 0xff;
-                posaccum = posaccum >> 8;
-            }
-        }
-        if (table_index < 6) {
-            for (uint64_t i = 0; i < right_writer_count; i++) {
-                globals.R_sort_manager->AddToCache(Bits(right_writer_buf + i * right_entry_size_bytes, right_entry_size_bytes,
-                     right_entry_size_bytes * 8));
-            }
-        } else {
-            // Writes out the right table for table 7
-            (*ptmp_1_disks)[table_index + 1].Write(
-                globals.right_writer, right_writer_buf, right_writer_count * right_entry_size_bytes);
-        }
-        globals.right_writer += right_writer_count * right_entry_size_bytes;
-        globals.right_writer_count += right_writer_count;
-
-        (*ptmp_1_disks)[table_index].Write(
-            globals.left_writer, left_writer_buf, left_writer_count * compressed_entry_size_bytes);
-        globals.left_writer += left_writer_count * compressed_entry_size_bytes;
-        globals.left_writer_count += left_writer_count;
-
-        globals.matches += matches;
-        SemaphoreUtils::Post(ptd->mine);
-    }
-
-    delete[] L_position_map;
-    delete[] R_position_map;
-
-    delete[] right_writer_buf;
-    delete[] left_writer_buf;
-    return 0;
-}
 
 class DiskPlotter {
 public:
@@ -608,24 +63,12 @@ public:
         uint32_t memo_len,
         const uint8_t* id,
         uint32_t id_len,
-        uint32_t buffmegabytes = 0,
-        uint32_t num_buckets = 0,
-        uint64_t stripe_size = 0,
-        uint8_t num_threads = 0)
+        uint32_t buf_megabytes_input = 0,
+        uint32_t num_buckets_input = 0,
+        uint64_t stripe_size_input = 0,
+        uint8_t num_threads_input = 0)
     {
-        if (stripe_size != 0) {
-            globals.stripe_size = stripe_size;
-        } else {
-            globals.stripe_size =  8192;
-        }
-        if (num_threads != 0) {
-            globals.num_threads = num_threads;
-        } else {
-            globals.num_threads = 2;
-        }
-        if (buffmegabytes == 0) {
-            buffmegabytes = 3500;
-        }
+        // Increases the open file limit, we will open a lot of files.
 #ifndef _WIN32
         struct rlimit the_limit = { 600, 600 };
         if (-1 == setrlimit(RLIMIT_NOFILE, &the_limit)) {
@@ -637,42 +80,63 @@ public:
             return;
         }
 
-        if (buffmegabytes < 10) {
+        uint32_t stripe_size, buf_megabytes, num_buckets;
+        uint8_t num_threads;
+        if (stripe_size_input != 0) {
+            stripe_size = stripe_size_input;
+        } else {
+            stripe_size =  8192;
+        }
+        if (num_threads_input != 0) {
+            num_threads = num_threads_input;
+        } else {
+            num_threads = 2;
+        }
+        if (buf_megabytes_input == 0) {
+            buf_megabytes = 3500;
+        }
+        if (num_threads == 0) {
+            std::cout << "Threads must be at least 1" << std::endl;
+            exit(1);
+        }
+
+        if (buf_megabytes < 10) {
             std::cout << "Please provide at least 10MiB of ram." << std::endl;
             exit(1);
         }
+
         // Subtract some ram to account for dynamic allocation through the code
-        uint64_t submbytes = (5 + (int)min(buffmegabytes * 0.05, (double)50));
-        memorySize = ((uint64_t)(buffmegabytes - submbytes)) * 1024 * 1024;
+        uint64_t submbytes = (5 + (int)min(buf_megabytes * 0.05, (double)50));
+        uint64_t memory_size = ((uint64_t)(buf_megabytes - submbytes)) * 1024 * 1024;
         double  max_table_size = 0;
         for (size_t i = 1; i <= 7; i++) {
-            double memory_i = 1.1 * ((uint64_t)1 << k) * GetMaxEntrySize(k, i, true);
+            double memory_i = 1.1 * ((uint64_t)1 << k) * EntrySizes::GetMaxEntrySize(k, i, true);
             if (memory_i > max_table_size) max_table_size = memory_i;
         }
-        if (num_buckets != 0){
-            this->numBuckets = Util::RoundPow2(num_buckets);
+        if (num_buckets_input != 0){
+            num_buckets = Util::RoundPow2(num_buckets_input);
         } else {
-            this->numBuckets = 2 * Util::RoundPow2(ceil(((double) max_table_size) / (memorySize * kMemSortProportion)));
+            num_buckets = 2 * Util::RoundPow2(ceil(((double) max_table_size) / (memory_size * kMemSortProportion)));
         }
 
-        if (this->numBuckets < kMinBuckets) {
-            if (num_buckets != 0) {
+        if (num_buckets < kMinBuckets) {
+            if (num_buckets_input != 0) {
                 std::cout << "Minimum buckets is " << kMinBuckets << std::endl;
                 exit(1);
             }
-            this->numBuckets = kMinBuckets;
-        } else if (this->numBuckets > kMaxBuckets) {
-            if (num_buckets != 0) {
+            num_buckets = kMinBuckets;
+        } else if (num_buckets > kMaxBuckets) {
+            if (num_buckets_input != 0) {
                 std::cout << "Maximum buckets is " << kMaxBuckets << std::endl;
                 exit(1);
             }
             std::cout << "Do not have enough memory. Need " << (max_table_size / kMaxBuckets) / kMemSortProportion / (1024 * 1024) + submbytes << " MiB" << std::endl;
             exit(1);
         }
-        this->logNumBuckets = log2(this->numBuckets);
-        assert(log2(this->numBuckets) == ceil(log2(this->numBuckets)));
+        uint32_t log_num_buckets = log2(num_buckets);
+        assert(log2(num_buckets) == ceil(log2(num_buckets)));
 
-        if (max_table_size / this->numBuckets < stripe_size * 30) {
+        if (max_table_size / num_buckets < stripe_size * 30) {
             std::cout << "Stripe size too large." << std::endl;
             exit(1);
         }
@@ -682,9 +146,9 @@ public:
                   << tmp2_dirname << std::endl;
         std::cout << "ID: " << Util::HexStr(id, id_len) << std::endl;
         std::cout << "Plot size is: " << static_cast<int>(k) << std::endl;
-        std::cout << "Buffer size is: " << buffmegabytes << "MiB" << std::endl;
-        std::cout << "Using " << this->numBuckets << " buckets" << std::endl;
-        std::cout << "Using " << (int)globals.num_threads << " threads of stripe size " << globals.stripe_size << std::endl;
+        std::cout << "Buffer size is: " << buf_megabytes << "MiB" << std::endl;
+        std::cout << "Using " << num_buckets << " buckets" << std::endl;
+        std::cout << "Using " << (int)num_threads << " threads of stripe size " << stripe_size << std::endl;
 
         // Cross platform way to concatenate paths, gulrak library.
         std::vector<fs::path> tmp_1_filenames = std::vector<fs::path>();
@@ -727,12 +191,11 @@ public:
         {
             // Scope for FileDisk
             std::vector<FileDisk> tmp_1_disks;
-            tmp_1_disks.push_back(FileDisk(tmp_1_filenames[0]));  //, (uint64_t)0));
+            tmp_1_disks.push_back(FileDisk(tmp_1_filenames[0]));
             for (int i = 1; i < 8; i++) {
                 uint64_t tablesize = ((uint64_t)1) << (k + 1);
-                tablesize = tablesize * GetMaxEntrySize(k, i, true);
-                cout << "tablesize " << i << " " << tablesize << endl;
-                tmp_1_disks.push_back(FileDisk(tmp_1_filenames[i]));  //, tablesize));
+                tablesize = tablesize * EntrySizes::GetMaxEntrySize(k, i, true);
+                tmp_1_disks.push_back(FileDisk(tmp_1_filenames[i]));
             }
 
             FileDisk tmp2_disk(tmp_2_filename);
@@ -741,16 +204,11 @@ public:
                 return;
             }
 
-            // These variables are used in the WriteParkToFile method. They are preallocatted here
-            // to save time.
-            parkToFileBytesSize = CalculateLinePointSize(k) + CalculateStubsSize(k) + 2 +
-                                  CalculateMaxDeltasSize(k, 1);
-            parkToFileBytes = new uint8_t[parkToFileBytesSize];
 
             assert(id_len == kIdLen);
 
             // Memory to be used for sorting and buffers
-            uint8_t* memory = new uint8_t[memorySize]();
+            uint8_t* memory = new uint8_t[memory_size]();
 
             std::cout << std::endl
                       << "Starting phase 1/4: Forward Propagation into tmp files... "
@@ -759,7 +217,7 @@ public:
             Timer p1;
             Timer all_phases;
             std::vector<uint64_t> table_sizes =
-                WritePlotFile(memory, tmp_1_disks, k, id, tmp_dirname, filename);
+                RunPhase1(memory, tmp_1_disks, k, id, tmp_dirname, filename, memory_size, num_buckets, log_num_buckets, stripe_size, num_threads);
             p1.PrintElapsed("Time for phase 1 =");
 
             std::cout << std::endl
@@ -768,22 +226,25 @@ public:
 
             Timer p2;
             std::vector<uint64_t> backprop_table_sizes =
-                Backpropagate(memory, tmp_1_disks, table_sizes, k, id, tmp_dirname, filename);
+                    RunPhase2(memory, tmp_1_disks, table_sizes, k, id, tmp_dirname, filename, memory_size, num_buckets, log_num_buckets);
             p2.PrintElapsed("Time for phase 2 =");
+
+            // Now we open a new file, where the final contents of the plot will be stored.
+            uint32_t header_size = WriteHeader(tmp2_disk, k, id, memo, memo_len);
 
             std::cout << std::endl
                       << "Starting phase 3/4: Compression from tmp files into " << tmp_2_filename
                       << " ... " << Timer::GetNow();
             Timer p3;
-            Phase3Results res = CompressTables(
-                memory, k, tmp2_disk, tmp_1_disks, backprop_table_sizes, id, memo, memo_len, tmp_dirname, filename);
+            Phase3Results res = RunPhase3(
+                memory, k, tmp2_disk, tmp_1_disks, backprop_table_sizes, id, tmp_dirname, filename, header_size, memory_size, num_buckets, log_num_buckets);
             p3.PrintElapsed("Time for phase 3 =");
 
             std::cout << std::endl
                       << "Starting phase 4/4: Write Checkpoint tables into " << tmp_2_filename
                       << " ... " << Timer::GetNow();
             Timer p4;
-            WriteCTables(k, k + 1, tmp2_disk, res);
+            RunPhase4(k, k + 1, tmp2_disk, res);
             p4.PrintElapsed("Time for phase 4 =");
 
             // The total number of bytes used for sort is saved to table_sizes[0]. All other
@@ -793,7 +254,7 @@ public:
             // estimate is conservative (high).
             uint64_t total_working_space = table_sizes[0];
             for (size_t i = 1; i <= 7; i++) {
-                total_working_space += table_sizes[i] * GetMaxEntrySize(k, i, false);
+                total_working_space += table_sizes[i] * EntrySizes::GetMaxEntrySize(k, i, false);
             }
             std::cout << "Approximate working space used (without final file): "
                       << static_cast<double>(total_working_space) / (1024 * 1024 * 1024) << " GiB"
@@ -806,7 +267,6 @@ public:
             all_phases.PrintElapsed("Total time =");
 
             delete[] memory;
-            delete[] parkToFileBytes;
         }
 
         std::cin.tie(prevstr);
@@ -873,90 +333,7 @@ public:
         } while (!bRenamed);
     }
 
-    static uint32_t GetMaxEntrySize(uint8_t k, uint8_t table_index, bool phase_1_size)
-    {
-        // This represents the largest entry size that each table will have, throughout the
-        // entire plotting process. This is useful because it allows us to rewrite tables
-        // on top of themselves without running out of space.
-        switch (table_index) {
-            case 1:
-                // Represents f1, x
-                if (phase_1_size) {
-                    return Util::ByteAlign(k + kExtraBits + k) / 8;
-                } else {
-                    // After computing matches, table 1 is rewritten without the f1, which
-                    // is useless after phase1.
-                    return Util::ByteAlign(k) / 8;
-                }
-            case 2:
-            case 3:
-            case 4:
-            case 5:
-            case 6:
-                if (phase_1_size)
-                    // If we are in phase 1, use the max size, with metadata.
-                    // Represents f, pos, offset, and metadata
-                    return Util::ByteAlign(
-                               k + kExtraBits + (k) + kOffsetSize +
-                               k * kVectorLens[table_index + 1]) /
-                           8;
-                else
-                    // If we are past phase 1, we can use a smaller size, the smaller between
-                    // phases 2 and 3. Represents either:
-                    //    a:  sort_key, pos, offset        or
-                    //    b:  line_point, sort_key
-                    return Util::ByteAlign(
-                               max(static_cast<uint32_t>(k + 1 + (k) + kOffsetSize),
-                                   static_cast<uint32_t>(2 * k + k + 1))) /
-                           8;
-            case 7:
-            default:
-                // Represents line_point, f7
-                return Util::ByteAlign(3 * k) / 8;
-        }
-    }
-
-    // Calculates the size of one C3 park. This will store bits for each f7 between
-    // two C1 checkpoints, depending on how many times that f7 is present. For low
-    // values of k, we need extra space to account for the additional variability.
-    static uint32_t CalculateC3Size(uint8_t k)
-    {
-        if (k < 20) {
-            return Util::ByteAlign(8 * kCheckpoint1Interval) / 8;
-        } else {
-            return Util::ByteAlign(kC3BitsPerEntry * kCheckpoint1Interval) / 8;
-        }
-    }
-
-    static uint32_t CalculateLinePointSize(uint8_t k) { return Util::ByteAlign(2 * k) / 8; }
-
-    // This is the full size of the deltas section in a park. However, it will not be fully filled
-    static uint32_t CalculateMaxDeltasSize(uint8_t k, uint8_t table_index)
-    {
-        if (table_index == 1) {
-            return Util::ByteAlign((kEntriesPerPark - 1) * kMaxAverageDeltaTable1) / 8;
-        }
-        return Util::ByteAlign((kEntriesPerPark - 1) * kMaxAverageDelta) / 8;
-    }
-
-    static uint32_t CalculateStubsSize(uint32_t k)
-    {
-        return Util::ByteAlign((kEntriesPerPark - 1) * (k - kStubMinusBits)) / 8;
-    }
-
-    static uint32_t CalculateParkSize(uint8_t k, uint8_t table_index)
-    {
-        return CalculateLinePointSize(k) + CalculateStubsSize(k) +
-               CalculateMaxDeltasSize(k, table_index);
-    }
-
 private:
-    uint64_t numBuckets;
-    uint64_t logNumBuckets;
-    uint64_t memorySize;
-    uint8_t* parkToFileBytes;
-    uint32_t parkToFileBytesSize;
-
     // Writes the plot file header to a file
     uint32_t WriteHeader(
         FileDisk& plot_Disk,
@@ -1007,1295 +384,6 @@ private:
             header_text.size() + kIdLen + 1 + 2 + kFormatDescription.size() + 2 + memo_len + 10 * 8;
         std::cout << "Wrote: " << bytes_written << std::endl;
         return bytes_written;
-    }
-
-    // This is Phase 1, or forward propagation. During this phase, all of the 7 tables,
-    // and f functions, are evaluated. The result is an intermediate plot file, that is
-    // several times larger than what the final file will be, but that has all of the
-    // proofs of space in it. First, F1 is computed, which is special since it uses
-    // ChaCha8, and each encryption provides multiple output values. Then, the rest of the
-    // f functions are computed, and a sort on disk happens for each table.
-    std::vector<uint64_t> WritePlotFile(
-        uint8_t* memory,
-        std::vector<FileDisk>& tmp_1_disks,
-        uint8_t k,
-        const uint8_t* id,
-        std::string tmp_dirname,
-        std::string filename)
-    {
-        std::cout << "Computing table 1" << std::endl;
-        Timer f1_start_time;
-        F1Calculator f1(k, id);
-        uint64_t x = 0;
-        uint64_t prevtableentries = 0;
-
-        uint32_t t1_entry_size_bytes = GetMaxEntrySize(k, 1, true);
-        globals.L_sort_manager = new LazySortManager(
-                memory,
-                memorySize,
-                this->numBuckets,
-                this->logNumBuckets,
-                t1_entry_size_bytes,
-                tmp_dirname,
-                filename + ".p1.t1",
-                0, globals.stripe_size);
-
-        // The max value our input (x), can take. A proof of space is 64 of these x values.
-        uint64_t max_value = ((uint64_t)1 << (k)) - 1;
-
-        // These are used for sorting on disk. The sort on disk code needs to know how
-        // many elements are in each bucket.
-        std::vector<uint64_t> bucket_sizes(kNumSortBuckets, 0);
-        std::vector<uint64_t> table_sizes = std::vector<uint64_t>(8, 0);
-
-        // Instead of computing f1(1), f1(2), etc, for each x, we compute them in batches
-        // to increase CPU efficency.
-        for (uint64_t lp = 0; lp <= (((uint64_t)1) << (k - kBatchSizes)); lp++) {
-            // For each pair x, y in the batch
-            for (auto kv : f1.CalculateBuckets(Bits(x, k), 2 << (kBatchSizes - 1))) {
-                Bits to_write = std::get<0>(kv) + std::get<1>(kv);
-                globals.L_sort_manager->AddToCache(to_write);
-                prevtableentries++;
-
-                if (x + 1 > max_value) {
-                    break;
-                }
-                ++x;
-            }
-            if (x + 1 > max_value) {
-                break;
-            }
-        }
-        f1_start_time.PrintElapsed("F1 complete, time:");
-        globals.L_sort_manager->FlushCache();
-        table_sizes[1] = x + 1;
-
-        double num_buckets = (((uint64_t)1) << (k + kExtraBits)) / static_cast<double>(kBC) + 1;
-
-        // Store positions to previous tables, in k bits.
-        uint8_t pos_size = k;
-        uint32_t right_entry_size_bytes = 0;
-
-        // For tables 1 through 6, sort the table, calculate matches, and write
-        // the next table. This is the left table index.
-        for (uint8_t table_index = 1; table_index < 7; table_index++) {
-            Timer table_timer;
-            uint8_t metadata_size = kVectorLens[table_index + 1] * k;
-
-            // Determines how many bytes the entries in our left and right tables will take up.
-            uint32_t entry_size_bytes = GetMaxEntrySize(k, table_index, true);
-            uint32_t compressed_entry_size_bytes = GetMaxEntrySize(k, table_index, false);
-            right_entry_size_bytes = GetMaxEntrySize(k, table_index + 1, true);
-
-            std::cout << "Computing table " << int{table_index + 1} << std::endl;
-            // Start of parallel execution
-
-            FxCalculator f(k, table_index + 1);  // dummy to load static table
-
-            globals.matches = 0;
-            // The memory will be used like this, with most memory allocated towards the SortManager, since it needs
-            // to sort large amounts of data.
-            // [----------------------------LSM/LR---------------------------|-----RSM/RW------|--LW--]
-            globals.left_reader_buf_size = floor(kMemSortProportion * memorySize);
-            globals.right_writer_buf_size =  3 * (memorySize - globals.left_reader_buf_size) / 4;
-            globals.left_writer_buf_size = (memorySize - globals.left_reader_buf_size - globals.right_writer_buf_size);
-
-            globals.left_reader_buf = &(memory[0]);
-            uint8_t* right_writer_buf = &(memory[globals.left_reader_buf_size]);
-            globals.left_writer_buf =  &(memory[globals.left_reader_buf_size + globals.right_writer_buf_size]);
-
-            globals.right_writer_buf_entries = globals.right_writer_buf_size / right_entry_size_bytes;
-            globals.left_writer_count = 0;
-            globals.right_writer_count = 0;
-            globals.right_writer = 0;
-            globals.left_writer = 0;
-
-            // Assign the first 3/4 of memory to the left reader (which contains the data from the previous iteration),
-            // And will need to perform large sorts.
-            globals.L_sort_manager->ChangeMemory(globals.left_reader_buf, globals.left_reader_buf_size);
-
-            globals.R_sort_manager = new LazySortManager(
-                    right_writer_buf,
-                    globals.right_writer_buf_size,
-                    this->numBuckets,
-                    this->logNumBuckets,
-                    right_entry_size_bytes,
-                    tmp_dirname,
-                    filename + ".p1.t" + to_string(table_index + 1),
-                    0,
-                    globals.stripe_size);
-
-            globals.L_sort_manager->TriggerNewBucket(0, 0);
-
-            Timer computation_pass_timer;
-
-            THREADDATA td[globals.num_threads];
-#ifdef _WIN32
-            HANDLE t[globals.num_threads];
-            HANDLE mutex[globals.num_threads];
-#else
-            pthread_t t[globals.num_threads];
-            sem_t* mutex[globals.num_threads];
-            char semname[20];
-#endif
-
-            for (int i = 0; i < globals.num_threads; i++) {
-#ifdef _WIN32
-                mutex[i] = CreateSemaphore(
-                    NULL,   // default security attributes
-                    0,      // initial count
-                    1,      // maximum count
-                    NULL);  // unnamed semaphore
-#else
-                sprintf(semname, "sem %d", i);
-                mutex[i] = sem_open(semname, O_CREAT, S_IRUSR | S_IWUSR, 0);
-#endif
-            }
-
-            for (int i = 0; i < globals.num_threads; i++) {
-                td[i].index = i;
-                td[i].mine = mutex[i];
-                td[i].theirs = mutex[(globals.num_threads + i - 1) % globals.num_threads];
-
-                td[i].prevtableentries = prevtableentries;
-                td[i].right_entry_size_bytes = right_entry_size_bytes;
-                td[i].k = k;
-                td[i].table_index = table_index;
-                td[i].metadata_size = metadata_size;
-                td[i].entry_size_bytes = entry_size_bytes;
-                td[i].pos_size = pos_size;
-                td[i].compressed_entry_size_bytes = compressed_entry_size_bytes;
-                td[i].ptmp_1_disks = &tmp_1_disks;
-
-#ifdef _WIN32
-                t[i] = CreateThread(0, 0, thread, &(td[i]), 0, NULL);
-#else
-                pthread_create(&(t[i]), NULL, thread, &(td[i]));
-#endif
-            }
-            SemaphoreUtils::Post(mutex[globals.num_threads - 1]);
-
-            for (int i = 0; i < globals.num_threads; i++) {
-#ifdef _WIN32
-                WaitForSingleObject(t[i], INFINITE);
-                CloseHandle(t[i]);
-#else
-                pthread_join(t[i], NULL);
-#endif
-            }
-
-            for (int i = 0; i < globals.num_threads; i++) {
-#ifdef _WIN32
-                CloseHandle(mutex[i]);
-#else
-                sem_close(mutex[i]);
-                sem_unlink(semname);
-#endif
-            }
-
-            // end of parallel execution
-
-            // Total matches found in the left table
-            std::cout << "\tTotal matches: " << globals.matches
-                      << ". Per bucket: " << (globals.matches / num_buckets) << std::endl;
-
-            table_sizes[table_index] = globals.left_writer_count + 1;
-            table_sizes[table_index + 1] = globals.right_writer_count + 1;
-
-            // Finishes writing L table, and writes the 0 entry (EOT) for left table
-            // Also truncates the file after the final write position, deleting no longer useful
-            // working space
-
-            uint8_t* zero_buf =
-                new uint8_t[std::max(compressed_entry_size_bytes, right_entry_size_bytes)];
-            memset(zero_buf, 0x00, std::max(compressed_entry_size_bytes, right_entry_size_bytes));
-
-            tmp_1_disks[table_index].Write(globals.left_writer, zero_buf, compressed_entry_size_bytes);
-            globals.left_writer += compressed_entry_size_bytes;
-            tmp_1_disks[table_index].Truncate(globals.left_writer);
-            delete globals.L_sort_manager;
-            if (table_index < 6) {
-                globals.R_sort_manager->FlushCache();
-                globals.L_sort_manager = globals.R_sort_manager;
-            } else {
-                // Writes the 0 entry (EOT) for right table
-                tmp_1_disks[table_index + 1].Write(globals.right_writer, zero_buf, right_entry_size_bytes);
-                globals.right_writer += right_entry_size_bytes;
-                tmp_1_disks[table_index + 1].Truncate(globals.right_writer);
-            }
-
-            // Resets variables
-            if (globals.matches != globals.right_writer_count) {
-                cout << "UNMATCHED!" << endl;
-                std::cout << globals.matches << " " << globals.right_writer_count << std::endl;
-                exit(1);
-            }
-
-            prevtableentries = globals.right_writer_count;
-            table_timer.PrintElapsed("Forward propagation table time:");
-
-            delete[] zero_buf;
-        }
-        table_sizes[0] = 0;
-        delete globals.R_sort_manager;
-        return table_sizes;
-    }
-
-    // Backpropagate takes in as input, a file on which forward propagation has been done.
-    // The purpose of backpropagate is to eliminate any dead entries that don't contribute
-    // to final values in f7, to minimize disk usage. A sort on disk is applied to each table,
-    // so that they are sorted by position.
-    std::vector<uint64_t> Backpropagate(
-            uint8_t *memory,
-            std::vector<FileDisk> &tmp_1_disks,
-            std::vector<uint64_t> table_sizes,
-            uint8_t k,
-            const uint8_t *id,
-            const std::string &tmp_dirname,
-            const std::string &filename)
-    {
-        // An extra bit is used, since we may have more than 2^k entries in a table. (After pruning,
-        // each table will have 0.8*2^k or less entries).
-        uint8_t pos_size = k;
-
-        std::vector<uint64_t> new_table_sizes = std::vector<uint64_t>(8, 0);
-        new_table_sizes[7] = table_sizes[7];
-        LazySortManager* R_sort_manager;
-        LazySortManager* L_sort_manager;
-
-        // Iterates through each table (with a left and right pointer), starting at 6 & 7.
-        for (uint8_t table_index = 7; table_index > 1; --table_index) {
-            // std::vector<std::pair<uint64_t, uint64_t> > match_positions;
-            Timer table_timer;
-
-            std::cout << "Backpropagating on table " << int{table_index} << std::endl;
-
-            uint16_t left_metadata_size = kVectorLens[table_index] * k;
-
-            // The entry that we are reading (no metadata)
-            uint16_t left_entry_size_bytes = GetMaxEntrySize(k, table_index - 1, false);
-
-            // The right entries which we read and write (the already have no metadata, since they
-            // have been pruned in previous iteration)
-            uint16_t right_entry_size_bytes = GetMaxEntrySize(k, table_index, false);
-
-            uint64_t left_reader = 0;
-            uint64_t left_writer = 0;
-            uint64_t right_reader = 0;
-            uint64_t right_writer = 0;
-            // The memory will be used like this, with most memory allocated towards the SortManager, since it needs it
-            // [--------------------------SM/RR-------------------------|-----------LW-------------|--RW--|--LR--]
-            uint64_t sort_manager_buf_size = floor(kMemSortProportion * memorySize);
-            uint64_t left_writer_buf_size = 3 * (memorySize - sort_manager_buf_size) / 4;
-            uint64_t other_buf_sizes = (memorySize - sort_manager_buf_size - left_writer_buf_size) / 2;
-            uint8_t *right_reader_buf = &(memory[0]);
-            uint8_t *left_writer_buf = &(memory[sort_manager_buf_size]);
-            uint8_t *right_writer_buf = &(memory[sort_manager_buf_size + left_writer_buf_size]);
-            uint8_t *left_reader_buf = &(memory[sort_manager_buf_size + left_writer_buf_size + other_buf_sizes]);
-            uint64_t right_reader_buf_entries = sort_manager_buf_size / right_entry_size_bytes;
-            uint64_t left_writer_buf_entries = left_writer_buf_size / left_entry_size_bytes;
-            uint64_t right_writer_buf_entries = other_buf_sizes / right_entry_size_bytes;
-            uint64_t left_reader_buf_entries = other_buf_sizes / left_entry_size_bytes;
-            uint64_t left_reader_count = 0;
-            uint64_t right_reader_count = 0;
-            uint64_t left_writer_count = 0;
-            uint64_t right_writer_count = 0;
-
-            if (table_index != 7) {
-                R_sort_manager->ChangeMemory(memory, sort_manager_buf_size);
-            }
-
-            L_sort_manager = new LazySortManager(
-                    left_writer_buf,
-                    left_writer_buf_size,
-                    this->numBuckets,
-                    this->logNumBuckets,
-                    left_entry_size_bytes,
-                    tmp_dirname,
-                    filename + ".p2.t" + to_string(table_index - 1),
-                    0,
-                    0);
-
-            // We will divide by 2, so it must be even.
-            assert(kCachedPositionsSize % 2 == 0);
-
-            // Used positions will be used to mark which posL are present in table R, the rest will
-            // be pruned
-            bool used_positions[kCachedPositionsSize];
-            memset(used_positions, 0, sizeof(used_positions));
-
-            bool should_read_entry = true;
-
-            // Cache for when we read a right entry that is too far forward
-            uint64_t cached_entry_sort_key = 0;  // For table_index == 7, y is here
-            uint64_t cached_entry_pos = 0;
-            uint64_t cached_entry_offset = 0;
-
-            uint64_t left_entry_counter = 0;  // Total left entries written
-
-            // Sliding window map, from old position to new position (after pruning)
-            uint64_t new_positions[kCachedPositionsSize];
-
-            // Sort keys represent the ordering of entries, sorted by (y, pos, offset),
-            // but using less bits (only k+1 instead of 2k + 9, etc.)
-            // This is a map from old position to array of sort keys (one for each R entry with this
-            // pos)
-            uint64_t old_sort_keys[kReadMinusWrite][kMaxMatchesSingleEntry];
-            // Map from old position to other positions that it matches with
-            uint64_t old_offsets[kReadMinusWrite][kMaxMatchesSingleEntry];
-            // Map from old position to count (number of times it appears)
-            uint16_t old_counters[kReadMinusWrite];
-
-            for (uint16_t &old_counter : old_counters) {
-                old_counter = 0;
-            }
-
-            bool end_of_right_table = false;
-            uint64_t current_pos =
-                    0;  // This is the current pos that we are looking for in the L table
-            uint64_t end_of_table_pos = 0;
-            uint64_t greatest_pos = 0;  // This is the greatest position we have seen in R table
-
-            // Buffers for reading and writing to disk
-            uint8_t *left_entry_buf;
-            uint8_t *new_left_entry_buf;
-            uint8_t *right_entry_buf;
-            uint8_t *right_entry_buf_SM = new uint8_t[right_entry_size_bytes];
-
-            // Go through all right entries, and keep going since write pointer is behind read
-            // pointer
-            while (!end_of_right_table || (current_pos - end_of_table_pos <= kReadMinusWrite)) {
-                old_counters[current_pos % kReadMinusWrite] = 0;
-
-                // Resets used positions after a while, so we use little memory
-                if ((current_pos - kReadMinusWrite) % (kCachedPositionsSize / 2) == 0) {
-                    if ((current_pos - kReadMinusWrite) % kCachedPositionsSize == 0) {
-                        for (uint32_t i = kCachedPositionsSize / 2; i < kCachedPositionsSize; i++) {
-                            used_positions[i] = false;
-                        }
-                    } else {
-                        for (uint32_t i = 0; i < kCachedPositionsSize / 2; i++) {
-                            used_positions[i] = false;
-                        }
-                    }
-                }
-                // Only runs this code if we are still reading the right table, or we still need to
-                // read more left table entries (current_pos <= greatest_pos), otherwise, it skips
-                // to the writing of the final R table entries
-                if (!end_of_right_table || current_pos <= greatest_pos) {
-                    uint64_t entry_sort_key = 0;
-                    uint64_t entry_pos = 0;
-                    uint64_t entry_offset = 0;
-
-                    while (!end_of_right_table) {
-                        if (should_read_entry) {
-                            if (right_reader_count == new_table_sizes[table_index] - 1) {
-                                // Table R has ended, don't read any more (but keep writing)
-                                end_of_right_table = true;
-                                end_of_table_pos = current_pos;
-                                break;
-                            }
-                            // Need to read another entry at the current position
-                            if (table_index == 7) {
-                                if (right_reader_count % right_reader_buf_entries == 0) {
-                                    uint64_t readAmt = std::min(
-                                            right_reader_buf_entries * right_entry_size_bytes,
-                                            (new_table_sizes[table_index] - right_reader_count) *
-                                            right_entry_size_bytes);
-
-                                    tmp_1_disks[table_index].Read(
-                                            right_reader, right_reader_buf, readAmt);
-                                    right_reader += readAmt;
-                                }
-                                right_entry_buf =
-                                        right_reader_buf +
-                                        (right_reader_count % right_reader_buf_entries) * right_entry_size_bytes;
-                            } else {
-                                right_entry_buf = R_sort_manager->ReadEntry(right_reader);
-                                right_reader += right_entry_size_bytes;
-                            }
-                            right_reader_count++;
-
-                            if (table_index == 7) {
-                                // This is actually y for table 7
-                                entry_sort_key = Util::SliceInt64FromBytes(right_entry_buf, 0, k);
-                                entry_pos = Util::SliceInt64FromBytes(right_entry_buf, k, pos_size);
-                                entry_offset = Util::SliceInt64FromBytes(
-                                        right_entry_buf, k + pos_size, kOffsetSize);
-                            } else {
-                                entry_pos = Util::SliceInt64FromBytes(right_entry_buf, 0, pos_size);
-                                entry_offset = Util::SliceInt64FromBytes(
-                                        right_entry_buf, pos_size, kOffsetSize);
-                                entry_sort_key = Util::SliceInt64FromBytes(
-                                        right_entry_buf, pos_size + kOffsetSize, k + 1);
-                            }
-                        } else if (cached_entry_pos == current_pos) {
-                            // We have a cached entry at this position
-                            entry_sort_key = cached_entry_sort_key;
-                            entry_pos = cached_entry_pos;
-                            entry_offset = cached_entry_offset;
-                        } else {
-                            // The cached entry is at a later pos, so we don't read any more R
-                            // entries, read more L entries instead.
-                            break;
-                        }
-
-                        should_read_entry = true;  // By default, read another entry
-                        if (entry_pos + entry_offset > greatest_pos) {
-                            // Greatest L pos that we should look for
-                            greatest_pos = entry_pos + entry_offset;
-                        }
-
-                        if (entry_pos == current_pos) {
-                            // The current L position is the current R entry
-                            // Marks the two matching entries as used (pos and pos+offset)
-                            used_positions[entry_pos % kCachedPositionsSize] = true;
-                            used_positions[(entry_pos + entry_offset) % kCachedPositionsSize] =
-                                    true;
-
-                            uint64_t old_write_pos = entry_pos % kReadMinusWrite;
-
-                            // Stores the sort key for this R entry
-                            old_sort_keys[old_write_pos][old_counters[old_write_pos]] =
-                                    entry_sort_key;
-
-                            // Stores the other matching pos for this R entry (pos6 + offset)
-                            old_offsets[old_write_pos][old_counters[old_write_pos]] =
-                                    entry_pos + entry_offset;
-                            ++old_counters[old_write_pos];
-                        } else {
-                            // Don't read any more right entries for now, because we haven't caught
-                            // up on the left table yet
-                            should_read_entry = false;
-                            cached_entry_sort_key = entry_sort_key;
-                            cached_entry_pos = entry_pos;
-                            cached_entry_offset = entry_offset;
-                            break;
-                        }
-                    }
-                    // ***Reads a left entry
-                    if (left_reader_count % left_reader_buf_entries == 0) {
-                        uint64_t readAmt = std::min(
-                                left_reader_buf_entries * left_entry_size_bytes,
-                                (table_sizes[table_index - 1] - left_reader_count) *
-                                left_entry_size_bytes);
-                        tmp_1_disks[table_index - 1].Read(left_reader, left_reader_buf, readAmt);
-                        left_reader += readAmt;
-                    }
-                    left_entry_buf = left_reader_buf +
-                                     (left_reader_count % left_reader_buf_entries) * left_entry_size_bytes;
-                    left_reader_count++;
-
-                    // If this left entry is used, we rewrite it. If it's not used, we ignore it.
-                    if (used_positions[current_pos % kCachedPositionsSize]) {
-                        uint64_t entry_metadata;
-
-                        if (table_index > 2) {
-                            // For tables 2-6, the entry is: pos, offset
-                            entry_pos = Util::SliceInt64FromBytes(left_entry_buf, 0, pos_size);
-                            entry_offset =
-                                    Util::SliceInt64FromBytes(left_entry_buf, pos_size, kOffsetSize);
-                        } else {
-                            entry_metadata =
-                                    Util::SliceInt64FromBytes(left_entry_buf, 0, left_metadata_size);
-                        }
-
-                        new_left_entry_buf =
-                                left_writer_buf +
-                                (left_writer_count % left_writer_buf_entries) * left_entry_size_bytes;
-                        left_writer_count++;
-
-                        Bits new_left_entry;
-                        if (table_index > 2) {
-                            // The new left entry is slightly different. Metadata is dropped, to
-                            // save space, and the counter of the entry is written (sort_key). We
-                            // use this instead of (y + pos + offset) since its smaller.
-                            new_left_entry += Bits(entry_pos, pos_size);
-                            new_left_entry += Bits(entry_offset, kOffsetSize);
-                            new_left_entry += Bits(left_entry_counter, k + 1);
-
-                            // If we are not taking up all the bits, make sure they are zeroed
-                            if (Util::ByteAlign(new_left_entry.GetSize()) <
-                                left_entry_size_bytes * 8) {
-                                new_left_entry +=
-                                        Bits(0, left_entry_size_bytes * 8 - new_left_entry.GetSize());
-                            }
-                            L_sort_manager->AddToCache(new_left_entry);
-                        } else {
-                            // For table one entries, we don't care about sort key, only x.
-                            // Also, we don't use the sort manager, since we won't sort it.
-                            new_left_entry += Bits(entry_metadata, left_metadata_size);
-                            new_left_entry.ToBytes(new_left_entry_buf);
-                            if (left_writer_count % left_writer_buf_entries == 0) {
-                                tmp_1_disks[table_index - 1].Write(
-                                        left_writer,
-                                        left_writer_buf,
-                                        left_writer_buf_entries * left_entry_size_bytes);
-                                left_writer += left_writer_buf_entries * left_entry_size_bytes;
-                            }
-                        }
-
-                        // Mapped positions, so we can rewrite the R entry properly
-                        new_positions[current_pos % kCachedPositionsSize] = left_entry_counter;
-
-                        // Counter for new left entries written
-                        ++left_entry_counter;
-                    }
-                }
-                // Write pointer lags behind the read pointer
-                int64_t write_pointer_pos = current_pos - kReadMinusWrite + 1;
-
-                // Only write entries for write_pointer_pos, if we are above 0, and there are
-                // actually R entries for that pos.
-                if (write_pointer_pos >= 0 &&
-                    used_positions[write_pointer_pos % kCachedPositionsSize]) {
-                    uint64_t new_pos = new_positions[write_pointer_pos % kCachedPositionsSize];
-                    Bits new_pos_bin(new_pos, pos_size);
-                    // There may be multiple R entries that share the write_pointer_pos, so write
-                    // all of them
-                    for (uint32_t counter = 0;
-                         counter < old_counters[write_pointer_pos % kReadMinusWrite];
-                         counter++) {
-                        // Creates and writes the new right entry, with the cached data
-                        uint64_t new_offset_pos = new_positions
-                        [old_offsets[write_pointer_pos % kReadMinusWrite][counter] %
-                         kCachedPositionsSize];
-                        Bits new_right_entry =
-                                table_index == 7
-                                ? Bits(
-                                        old_sort_keys[write_pointer_pos % kReadMinusWrite][counter],
-                                        k)
-                                : Bits(
-                                        old_sort_keys[write_pointer_pos % kReadMinusWrite][counter],
-                                        k + 1);
-                        new_right_entry += new_pos_bin;
-                        // match_positions.push_back(std::make_pair(new_pos, new_offset_pos));
-                        new_right_entry.AppendValue(new_offset_pos - new_pos, kOffsetSize);
-
-                        // Calculate right entry pointer for output
-                        right_entry_buf =
-                                right_writer_buf +
-                                (right_writer_count % right_writer_buf_entries) * right_entry_size_bytes;
-                        right_writer_count++;
-
-                        if (Util::ByteAlign(new_right_entry.GetSize()) <
-                            right_entry_size_bytes * 8) {
-                            memset(right_entry_buf, 0, right_entry_size_bytes);
-                        }
-                        new_right_entry.ToBytes(right_entry_buf);
-                        // Check for write out to disk
-                        if (right_writer_count % right_writer_buf_entries == 0) {
-                            tmp_1_disks[table_index].Write(
-                                    right_writer,
-                                    right_writer_buf,
-                                    right_writer_buf_entries * right_entry_size_bytes);
-                            right_writer += right_writer_buf_entries * right_entry_size_bytes;
-                        }
-                    }
-                }
-                ++current_pos;
-            }
-            new_table_sizes[table_index - 1] = left_entry_counter + 1;
-
-            std::cout << "\tWrote left entries: " << left_entry_counter << std::endl;
-            table_timer.PrintElapsed("Total backpropagation time::");
-
-            tmp_1_disks[table_index].Write(
-                    right_writer,
-                    right_writer_buf,
-                    (right_writer_count % right_writer_buf_entries) * right_entry_size_bytes);
-            right_writer += (right_writer_count % right_writer_buf_entries) * right_entry_size_bytes;
-
-            if (table_index != 7) {
-                delete R_sort_manager;
-            }
-
-            // Writes the 0 entry (EOT) for right table
-            memset(right_writer_buf, 0x00, right_entry_size_bytes);
-            tmp_1_disks[table_index].Write(right_writer, right_writer_buf, right_entry_size_bytes);
-            right_writer += right_entry_size_bytes;
-            tmp_1_disks[table_index].Truncate(right_writer);
-
-            if (table_index == 2) {
-                // Writes remaining entries for table1
-                tmp_1_disks[table_index - 1].Write(
-                        left_writer,
-                        left_writer_buf,
-                        (left_writer_count % left_writer_buf_entries) * left_entry_size_bytes);
-                left_writer += (left_writer_count % left_writer_buf_entries) * left_entry_size_bytes;
-
-                // Writes the 0 entry (EOT) for left table
-                memset(left_writer_buf, 0x00, left_entry_size_bytes);
-                tmp_1_disks[table_index - 1].Write(left_writer, left_writer_buf, left_entry_size_bytes);
-                left_writer += left_entry_size_bytes;
-                tmp_1_disks[table_index - 1].Truncate(left_writer);
-            } else {
-                L_sort_manager->FlushCache();
-                R_sort_manager = L_sort_manager;
-            }
-            delete[] right_entry_buf_SM;
-        }
-        delete L_sort_manager;
-        return new_table_sizes;
-    }
-
-    // This writes a number of entries into a file, in the final, optimized format. The park
-    // contains a checkpoint value (which is a 2k bits line point), as well as EPP (entries per
-    // park) entries. These entries are each divided into stub and delta section. The stub bits are
-    // encoded as is, but the delta bits are optimized into a variable encoding scheme. Since we
-    // have many entries in each park, we can approximate how much space each park with take. Format
-    // is: [2k bits of first_line_point]  [EPP-1 stubs] [Deltas size] [EPP-1 deltas]....
-    // [first_line_point] ...
-    void WriteParkToFile(
-            FileDisk &final_disk,
-            uint64_t table_start,
-            uint64_t park_index,
-            uint32_t park_size_bytes,
-            uint128_t first_line_point,
-            const std::vector<uint8_t> &park_deltas,
-            const std::vector<uint64_t> &park_stubs,
-            uint8_t k,
-            uint8_t table_index)
-    {
-        // Parks are fixed size, so we know where to start writing. The deltas will not go over
-        // into the next park.
-        uint64_t writer = table_start + park_index * park_size_bytes;
-        uint8_t *index = parkToFileBytes;
-
-        Bits first_line_point_bits(first_line_point, 2 * k);
-        first_line_point_bits.ToBytes(index);
-        index += CalculateLinePointSize(k);
-
-        // We use ParkBits instead of Bits since it allows storing more data
-        ParkBits park_stubs_bits;
-        for (uint64_t stub : park_stubs) {
-            park_stubs_bits.AppendValue(stub, (k - kStubMinusBits));
-        }
-        uint32_t stubs_size = CalculateStubsSize(k);
-        uint32_t stubs_valid_size = CDIV(park_stubs_bits.GetSize(), 8);
-        park_stubs_bits.ToBytes(index);
-        memset(index + stubs_valid_size, 0, stubs_size - stubs_valid_size);
-        index += stubs_size;
-
-        // The stubs are random so they don't need encoding. But deltas are more likely to
-        // be small, so we can compress them
-        double R = kRValues[table_index - 1];
-        uint8_t *deltas_start = index + 2;
-        size_t deltas_size = Encoding::ANSEncodeDeltas(park_deltas, R, deltas_start);
-
-        if (!deltas_size) {
-            // Uncompressed
-            deltas_size = park_deltas.size();
-            Util::IntToTwoBytesLE(index, deltas_size | 0x8000);
-            memcpy(deltas_start, park_deltas.data(), deltas_size);
-        } else {
-            // Compressed
-            Util::IntToTwoBytesLE(index, deltas_size);
-        }
-
-        index += 2 + deltas_size;
-
-        memset(index, 0x00, park_size_bytes - (index - parkToFileBytes));
-        if ((uint32_t)(index - parkToFileBytes) > parkToFileBytesSize)
-            std::cout << "index-parkToFileBytes " << index - parkToFileBytes
-                      << " parkToFileBytesSize " << parkToFileBytesSize << std::endl;
-
-        final_disk.Write(writer, (uint8_t *)parkToFileBytes, park_size_bytes);
-    }
-
-    // Compresses the plot file tables into the final file. In order to do this, entries must be
-    // reorganized from the (pos, offset) bucket sorting order, to a more free line_point sorting
-    // order. In (pos, offset ordering), we store two pointers two the previous table, (x, y) which
-    // are very close together, by storing  (x, y-x), or (pos, offset), which can be done in about k
-    // + 8 bits, since y is in the next bucket as x. In order to decrease this, We store the actual
-    // entries from the previous table (e1, e2), instead of pos, offset pointers, and sort the
-    // entire table by (e1,e2). Then, the deltas between each (e1, e2) can be stored, which require
-    // around k bits.
-
-    // Converting into this format requires a few passes and sorts on disk. It also assumes that the
-    // backpropagation step happened, so there will be no more dropped entries. See the design
-    // document for more details on the algorithm.
-    Phase3Results CompressTables(
-            uint8_t *memory,
-            uint8_t k,
-            FileDisk &tmp2_disk /*filename*/,
-            std::vector<FileDisk> &tmp_1_disks /*plot_filename*/,
-            std::vector<uint64_t> table_sizes,
-            const uint8_t *id,
-            const uint8_t *memo,
-            uint32_t memo_len,
-            const std::string &tmp_dirname,
-            const std::string &filename)
-    {
-        // In this phase we open a new file, where the final contents of the plot will be stored.
-        uint32_t header_size = WriteHeader(tmp2_disk, k, id, memo, memo_len);
-
-        uint8_t pos_size = k;
-
-        std::vector<uint64_t> final_table_begin_pointers(12, 0);
-        final_table_begin_pointers[1] = header_size;
-
-        uint8_t table_pointer_bytes[8];
-        Util::IntToEightBytes(table_pointer_bytes, final_table_begin_pointers[1]);
-        tmp2_disk.Write(header_size - 10 * 8, table_pointer_bytes, 8);
-
-        uint64_t final_entries_written = 0;
-        uint32_t right_entry_size_bytes = 0;
-
-        LazySortManager* L_sort_manager;
-        LazySortManager* R_sort_manager;
-
-        // Iterates through all tables, starting at 1, with L and R pointers.
-        // For each table, R entries are rewritten with line points. Then, the right table is
-        // sorted by line_point. After this, the right table entries are rewritten as (sort_key,
-        // new_pos), where new_pos is the position in the table, where it's sorted by line_point,
-        // and the line_points are written to disk to a final table. Finally, table_i is sorted by
-        // sort_key. This allows us to compare to the next table.
-        for (uint8_t table_index = 1; table_index < 7; table_index++) {
-            Timer table_timer;
-            Timer computation_pass_1_timer;
-            std::cout << "Compressing tables " << int{table_index} << " and "
-                      << int{table_index + 1} << std::endl;
-
-            // The park size must be constant, for simplicity, but must be big enough to store EPP
-            // entries. entry deltas are encoded with variable length, and thus there is no
-            // guarantee that they won't override into the next park. It is only different (larger)
-            // for table 1
-            uint32_t park_size_bytes = CalculateParkSize(k, table_index);
-
-            // Sort key for table 7 is just y, which is k bits. For all other tables it can
-            // be higher than 2^k and therefore k+1 bits are used.
-            uint32_t right_sort_key_size = table_index == 6 ? k : k + 1;
-
-            uint32_t left_entry_size_bytes = GetMaxEntrySize(k, table_index, false);
-            right_entry_size_bytes = GetMaxEntrySize(k, table_index + 1, false);
-
-            uint64_t left_reader = 0;
-            uint64_t right_reader = 0;
-            // The memory will be used like this, with most memory allocated towards the SortManager, since it needs it
-            // [---------------------------SM/LR---------------------|----------RW--------|---RR---]
-            uint64_t sort_manager_buf_size = floor(kMemSortProportion * memorySize);
-            uint64_t right_writer_buf_size = 3 * (memorySize - sort_manager_buf_size) / 4;
-            uint64_t right_reader_buf_size = memorySize - sort_manager_buf_size - right_writer_buf_size;
-            uint8_t *left_reader_buf = &(memory[0]);
-            uint8_t *right_writer_buf = &(memory[sort_manager_buf_size]);
-            uint8_t *right_reader_buf = &(memory[sort_manager_buf_size + right_writer_buf_size]);
-            uint64_t left_reader_buf_entries = sort_manager_buf_size / left_entry_size_bytes;
-            uint64_t right_reader_buf_entries = right_reader_buf_size / right_entry_size_bytes;
-            uint64_t left_reader_count = 0;
-            uint64_t right_reader_count = 0;
-            uint64_t total_r_entries = 0;
-
-
-            if (table_index > 1) {
-                L_sort_manager->ChangeMemory(memory, sort_manager_buf_size);
-            }
-
-            R_sort_manager = new LazySortManager(
-                    right_writer_buf,
-                    right_writer_buf_size,
-                    this->numBuckets * 2,
-                    this->logNumBuckets + 1,
-                    right_entry_size_bytes,
-                    tmp_dirname,
-                    filename + ".p3.t" + to_string(table_index + 1),
-                    0, 0);
-
-            bool should_read_entry = true;
-            std::vector<uint64_t> left_new_pos(kCachedPositionsSize);
-
-            uint64_t old_sort_keys[kReadMinusWrite][kMaxMatchesSingleEntry];
-            uint64_t old_offsets[kReadMinusWrite][kMaxMatchesSingleEntry];
-            uint16_t old_counters[kReadMinusWrite];
-            for (uint16_t &old_counter : old_counters) {
-                old_counter = 0;
-            }
-            bool end_of_right_table = false;
-            uint64_t current_pos = 0;
-            uint64_t end_of_table_pos = 0;
-            uint64_t greatest_pos = 0;
-
-            uint8_t *right_entry_buf;
-            uint8_t *left_entry_disk_buf;
-            uint8_t *left_entry_buf_sm = new uint8_t[left_entry_size_bytes];
-
-            uint64_t entry_sort_key, entry_pos, entry_offset;
-            uint64_t cached_entry_sort_key = 0;
-            uint64_t cached_entry_pos = 0;
-            uint64_t cached_entry_offset = 0;
-
-            // Similar algorithm as Backprop, to read both L and R tables simultaneously
-            while (!end_of_right_table || (current_pos - end_of_table_pos <= kReadMinusWrite)) {
-                old_counters[current_pos % kReadMinusWrite] = 0;
-
-                if (end_of_right_table || current_pos <= greatest_pos) {
-                    while (!end_of_right_table) {
-                        if (should_read_entry) {
-                            if (right_reader_count == table_sizes[table_index + 1] - 1) {
-                                end_of_right_table = true;
-                                end_of_table_pos = current_pos;
-                                break;
-                            }
-                            // The right entries are in the format from backprop, (sort_key, pos,
-                            // offset)
-                            if (right_reader_count % right_reader_buf_entries == 0) {
-                                uint64_t readAmt = std::min(
-                                        right_reader_buf_entries * right_entry_size_bytes,
-                                        (table_sizes[table_index + 1] - right_reader_count) *
-                                        right_entry_size_bytes);
-
-                                tmp_1_disks[table_index + 1].Read(
-                                        right_reader, right_reader_buf, readAmt);
-                                right_reader += readAmt;
-                            }
-                            right_entry_buf =
-                                    right_reader_buf +
-                                    (right_reader_count % right_reader_buf_entries) * right_entry_size_bytes;
-                            right_reader_count++;
-
-                            entry_sort_key =
-                                    Util::SliceInt64FromBytes(right_entry_buf, 0, right_sort_key_size);
-                            entry_pos = Util::SliceInt64FromBytes(
-                                    right_entry_buf, right_sort_key_size, pos_size);
-                            entry_offset = Util::SliceInt64FromBytes(
-                                    right_entry_buf, right_sort_key_size + pos_size, kOffsetSize);
-                        } else if (cached_entry_pos == current_pos) {
-                            entry_sort_key = cached_entry_sort_key;
-                            entry_pos = cached_entry_pos;
-                            entry_offset = cached_entry_offset;
-                        } else {
-                            break;
-                        }
-
-                        should_read_entry = true;
-
-                        if (entry_pos + entry_offset > greatest_pos) {
-                            greatest_pos = entry_pos + entry_offset;
-                        }
-                        if (entry_pos == current_pos) {
-                            uint64_t old_write_pos = entry_pos % kReadMinusWrite;
-                            old_sort_keys[old_write_pos][old_counters[old_write_pos]] =
-                                    entry_sort_key;
-                            old_offsets[old_write_pos][old_counters[old_write_pos]] =
-                                    (entry_pos + entry_offset);
-                            ++old_counters[old_write_pos];
-                        } else {
-                            should_read_entry = false;
-                            cached_entry_sort_key = entry_sort_key;
-                            cached_entry_pos = entry_pos;
-                            cached_entry_offset = entry_offset;
-                            break;
-                        }
-                    }
-                    if (left_reader_count < table_sizes[table_index] - 1) {
-                        // The left entries are in the new format: (sort_key, new_pos), except for table
-                        // 1: (y, x).
-                        if (table_index == 1) {
-                            if (left_reader_count % left_reader_buf_entries == 0) {
-                                uint64_t readAmt = std::min(
-                                        left_reader_buf_entries * left_entry_size_bytes,
-                                        (table_sizes[table_index] - left_reader_count) * left_entry_size_bytes);
-
-                                tmp_1_disks[table_index].Read(left_reader, left_reader_buf, readAmt);
-                                left_reader += readAmt;
-                            }
-                            left_entry_disk_buf = left_reader_buf + (left_reader_count % left_reader_buf_entries) *
-                                                                    left_entry_size_bytes;
-                        } else {
-                            left_entry_disk_buf = L_sort_manager->ReadEntry(left_reader, 1);
-                            left_reader += left_entry_size_bytes;
-                        }
-                        left_reader_count++;
-                    }
-
-                    // We read the "new_pos" from the L table, which for table 1 is just x. For
-                    // other tables, the new_pos
-                    if (table_index == 1) {
-                        // Only k bits, since this is x
-                        left_new_pos[current_pos % kCachedPositionsSize] =
-                                Util::SliceInt64FromBytes(left_entry_disk_buf, 0, k);
-                    } else {
-                        // k+1 bits in case it overflows
-                        left_new_pos[current_pos % kCachedPositionsSize] =
-                                Util::SliceInt64FromBytes(left_entry_disk_buf, k + 1, k + 1);
-                    }
-                }
-
-                uint64_t write_pointer_pos = current_pos - kReadMinusWrite + 1;
-
-                // Rewrites each right entry as (line_point, sort_key)
-                if (current_pos + 1 >= kReadMinusWrite) {
-                    uint64_t left_new_pos_1 =
-                            left_new_pos[write_pointer_pos % kCachedPositionsSize];
-                    for (uint32_t counter = 0;
-                         counter < old_counters[write_pointer_pos % kReadMinusWrite];
-                         counter++) {
-                        uint64_t left_new_pos_2 = left_new_pos
-                        [old_offsets[write_pointer_pos % kReadMinusWrite][counter] %
-                         kCachedPositionsSize];
-
-                        // A line point is an encoding of two k bit values into one 2k bit value.
-                        uint128_t line_point =
-                                Encoding::SquareToLinePoint(left_new_pos_1, left_new_pos_2);
-
-                        if (left_new_pos_1 > ((uint64_t)1 << k) ||
-                            left_new_pos_2 > ((uint64_t)1 << k)) {
-                            std::cout << "left or right positions too large" << std::endl;
-                            std::cout << (line_point > ((uint128_t)1 << (2 * k)));
-                            if ((line_point > ((uint128_t)1 << (2 * k)))) {
-                                std::cout << "L, R: " << left_new_pos_1 << " " << left_new_pos_2
-                                          << std::endl;
-                                std::cout << "Line point: " << line_point << std::endl;
-                                abort();
-                            }
-                        }
-                        Bits to_write = Bits(line_point, 2 * k);
-                        to_write += Bits(
-                                old_sort_keys[write_pointer_pos % kReadMinusWrite][counter],
-                                right_sort_key_size);
-
-                        R_sort_manager->AddToCache(to_write);
-                        total_r_entries++;
-                    }
-                }
-                current_pos += 1;
-            }
-            computation_pass_1_timer.PrintElapsed("\tFirst computation pass time:");
-
-            // Remove no longer needed file
-            tmp_1_disks[table_index].Truncate(0);
-
-            // Flush cache so all entries are written to buckets
-            R_sort_manager->FlushCache();
-
-            delete[] left_entry_buf_sm;
-
-            Timer computation_pass_2_timer;
-
-            // The memory will be used like this, with most memory allocated towards the LeftSortManager, since it needs it
-            // [---------------------------LSM/RR-----------------------------------|---------RSM/RW---------]
-            right_reader = 0;
-            right_reader_buf_size =  floor(kMemSortProportion * memorySize);
-            right_writer_buf_size =  memorySize - right_reader_buf_size;
-            right_reader_buf = &(memory[0]);
-            right_writer_buf = &(memory[right_reader_buf_size]);
-            right_reader_count = 0;
-            uint64_t final_table_writer = final_table_begin_pointers[table_index];
-
-            final_entries_written = 0;
-
-            if (table_index > 1) {
-                // Make sure all files are removed
-                delete L_sort_manager;
-            }
-
-            // L sort manager will be used for the writer, and R sort manager will be used for the reader
-            R_sort_manager->ChangeMemory(right_reader_buf, right_reader_buf_size);
-            L_sort_manager = new LazySortManager(
-                    right_writer_buf,
-                    right_writer_buf_size,
-                    this->numBuckets,
-                    this->logNumBuckets,
-                    right_entry_size_bytes,
-                    tmp_dirname,
-                    filename + ".p3s.t" + to_string(table_index + 1),
-                    0,
-                    0);
-
-            std::vector<uint8_t> park_deltas;
-            std::vector<uint64_t> park_stubs;
-            uint128_t checkpoint_line_point = 0;
-            uint128_t last_line_point = 0;
-            uint64_t park_index = 0;
-
-            uint8_t* right_reader_entry_buf;
-
-            // Now we will write on of the final tables, since we have a table sorted by line point.
-            // The final table will simply store the deltas between each line_point, in fixed space
-            // groups(parks), with a checkpoint in each group.
-            Bits right_entry_bits;
-            int added_to_cache = 0;
-            for (uint64_t index = 0; index < total_r_entries; index++) {
-                right_reader_entry_buf = R_sort_manager->ReadEntry(right_reader, 2);
-                right_reader += right_entry_size_bytes;
-                right_reader_count++;
-
-                // Right entry is read as (line_point, sort_key)
-                uint128_t line_point = Util::SliceInt128FromBytes(right_reader_entry_buf, 0, 2 * k);
-                uint64_t sort_key =
-                        Util::SliceInt64FromBytes(right_reader_entry_buf, 2 * k, right_sort_key_size);
-
-                // Write the new position (index) and the sort key
-                Bits to_write = Bits(sort_key, right_sort_key_size);
-                to_write += Bits(index, k + 1);
-
-                L_sort_manager->AddToCache(to_write);
-                added_to_cache++;
-
-                // Every EPP entries, writes a park
-                if (index % kEntriesPerPark == 0) {
-                    if (index != 0) {
-                        WriteParkToFile(
-                                tmp2_disk,
-                                final_table_begin_pointers[table_index],
-                                park_index,
-                                park_size_bytes,
-                                checkpoint_line_point,
-                                park_deltas,
-                                park_stubs,
-                                k,
-                                table_index);
-                        park_index += 1;
-                        final_entries_written += (park_stubs.size() + 1);
-                    }
-                    park_deltas.clear();
-                    park_stubs.clear();
-
-                    checkpoint_line_point = line_point;
-                }
-                uint128_t big_delta = line_point - last_line_point;
-
-                // Since we have approx 2^k line_points between 0 and 2^2k, the average
-                // space between them when sorted, is k bits. Much more efficient than storing each
-                // line point. This is diveded into the stub and delta. The stub is the least
-                // significant (k-kMinusStubs) bits, and largely random/incompressible. The small
-                // delta is the rest, which can be efficiently encoded since it's usually very
-                // small.
-
-                uint64_t stub = big_delta & ((1ULL << (k - kStubMinusBits)) - 1);
-                uint64_t small_delta = big_delta >> (k - kStubMinusBits);
-
-                assert(small_delta < 256);
-
-                if ((index % kEntriesPerPark != 0)) {
-                    park_deltas.push_back(small_delta);
-                    park_stubs.push_back(stub);
-                }
-                last_line_point = line_point;
-            }
-            delete R_sort_manager;
-            L_sort_manager->FlushCache();
-
-            computation_pass_2_timer.PrintElapsed("\tSecond computation pass time:");
-
-            if (park_deltas.size() > 0) {
-                // Since we don't have a perfect multiple of EPP entries, this writes the last ones
-                WriteParkToFile(
-                        tmp2_disk,
-                        final_table_begin_pointers[table_index],
-                        park_index,
-                        park_size_bytes,
-                        checkpoint_line_point,
-                        park_deltas,
-                        park_stubs,
-                        k,
-                        table_index);
-                final_entries_written += (park_stubs.size() + 1);
-            }
-
-            Encoding::ANSFree(kRValues[table_index - 1]);
-            std::cout << "\tWrote " << final_entries_written << " entries" << std::endl;
-
-            final_table_begin_pointers[table_index + 1] =
-                    final_table_begin_pointers[table_index] + (park_index + 1) * park_size_bytes;
-
-            final_table_writer = header_size - 8 * (10 - table_index);
-            Util::IntToEightBytes(table_pointer_bytes, final_table_begin_pointers[table_index + 1]);
-            tmp2_disk.Write(final_table_writer, (table_pointer_bytes), 8);
-            final_table_writer += 8;
-
-            table_timer.PrintElapsed("Total compress table time:");
-        }
-
-        L_sort_manager->ChangeMemory(memory, memorySize);
-
-        // These results will be used to write table P7 and the checkpoint tables in phase 4.
-        return Phase3Results{
-                final_table_begin_pointers,
-                final_entries_written,
-                right_entry_size_bytes * 8,
-                header_size,
-                L_sort_manager};
-    }
-
-    // Writes the checkpoint tables. The purpose of these tables, is to store a list of ~2^k values
-    // of size k (the proof of space outputs from table 7), in a way where they can be looked up for
-    // proofs, but also efficiently. To do this, we assume table 7 is sorted by f7, and we write the
-    // deltas between each f7 (which will be mostly 1s and 0s), with a variable encoding scheme
-    // (C3). Furthermore, we create C1 checkpoints along the way.  For example, every 10,000 f7
-    // entries, we can have a C1 checkpoint, and a C3 delta encoded entry with 10,000 deltas.
-
-    // Since we can't store all the checkpoints in
-    // memory for large plots, we create checkpoints for the checkpoints (C2), that are meant to be
-    // stored in memory during proving. For example, every 10,000 C1 entries, we can have a C2
-    // entry.
-
-    // The final table format for the checkpoints will be:
-    // C1 (checkpoint values)
-    // C2 (checkpoint values into)
-    // C3 (deltas of f7s between C1 checkpoints)
-    void WriteCTables(
-            uint8_t k,
-            uint8_t pos_size,
-            FileDisk &tmp2_disk /*filename*/,
-            Phase3Results &res)
-    {
-        uint32_t P7_park_size = Util::ByteAlign((k + 1) * kEntriesPerPark) / 8;
-        uint64_t number_of_p7_parks =
-                ((res.final_entries_written == 0 ? 0 : res.final_entries_written - 1) /
-                 kEntriesPerPark) +
-                1;
-
-        uint64_t begin_byte_C1 =
-                res.final_table_begin_pointers[7] + number_of_p7_parks * P7_park_size;
-
-        uint64_t total_C1_entries = CDIV(res.final_entries_written, kCheckpoint1Interval);
-        uint64_t begin_byte_C2 = begin_byte_C1 + (total_C1_entries + 1) * (Util::ByteAlign(k) / 8);
-        uint64_t total_C2_entries = CDIV(total_C1_entries, kCheckpoint2Interval);
-        uint64_t begin_byte_C3 = begin_byte_C2 + (total_C2_entries + 1) * (Util::ByteAlign(k) / 8);
-
-        uint32_t size_C3 = CalculateC3Size(k);
-        uint64_t end_byte = begin_byte_C3 + (total_C1_entries)*size_C3;
-
-        res.final_table_begin_pointers[8] = begin_byte_C1;
-        res.final_table_begin_pointers[9] = begin_byte_C2;
-        res.final_table_begin_pointers[10] = begin_byte_C3;
-        res.final_table_begin_pointers[11] = end_byte;
-
-        uint64_t plot_file_reader = 0;
-        uint64_t final_file_writer_1 = begin_byte_C1;
-        uint64_t final_file_writer_2 = begin_byte_C3;
-        uint64_t final_file_writer_3 = res.final_table_begin_pointers[7];
-
-        uint64_t prev_y = 0;
-        std::vector<Bits> C2;
-        uint64_t num_C1_entries = 0;
-        vector<uint8_t> deltas_to_write;
-        uint32_t right_entry_size_bytes = res.right_entry_size_bits / 8;
-
-        uint8_t* right_entry_buf;
-        auto C1_entry_buf = new uint8_t[Util::ByteAlign(k) / 8];
-        auto C3_entry_buf = new uint8_t[size_C3];
-        auto P7_entry_buf = new uint8_t[P7_park_size];
-
-        std::cout << "\tStarting to write C1 and C3 tables" << std::endl;
-
-        ParkBits to_write_p7;
-
-        // We read each table7 entry, which is sorted by f7, but we don't need f7 anymore. Instead,
-        // we will just store pos6, and the deltas in table C3, and checkpoints in tables C1 and C2.
-        for (uint64_t f7_position = 0; f7_position < res.final_entries_written; f7_position++) {
-            right_entry_buf = res.table7_sm->ReadEntry(plot_file_reader, 1);
-
-            plot_file_reader += right_entry_size_bytes;
-            uint64_t entry_y = Util::SliceInt64FromBytes(right_entry_buf, 0, k);
-            uint64_t entry_new_pos = Util::SliceInt64FromBytes(right_entry_buf, k, pos_size);
-
-            Bits entry_y_bits = Bits(entry_y, k);
-
-            if (f7_position % kEntriesPerPark == 0 && f7_position > 0) {
-                memset(P7_entry_buf, 0, P7_park_size);
-                to_write_p7.ToBytes(P7_entry_buf);
-                tmp2_disk.Write(final_file_writer_3, (P7_entry_buf), P7_park_size);
-                final_file_writer_3 += P7_park_size;
-                to_write_p7 = ParkBits();
-            }
-
-            to_write_p7 += ParkBits(entry_new_pos, k + 1);
-
-            if (f7_position % kCheckpoint1Interval == 0) {
-                entry_y_bits.ToBytes(C1_entry_buf);
-                tmp2_disk.Write(final_file_writer_1, (C1_entry_buf), Util::ByteAlign(k) / 8);
-                final_file_writer_1 += Util::ByteAlign(k) / 8;
-                if (num_C1_entries > 0) {
-                    final_file_writer_2 = begin_byte_C3 + (num_C1_entries - 1) * size_C3;
-                    size_t num_bytes =
-                            Encoding::ANSEncodeDeltas(deltas_to_write, kC3R, C3_entry_buf + 2) + 2;
-
-                    // We need to be careful because deltas are variable sized, and they need to fit
-                    assert(size_C3 * 8 > num_bytes);
-
-                    // Write the size
-                    Util::IntToTwoBytes(C3_entry_buf, num_bytes - 2);
-
-                    tmp2_disk.Write(final_file_writer_2, (C3_entry_buf), num_bytes);
-                    final_file_writer_2 += num_bytes;
-                }
-                prev_y = entry_y;
-                if (f7_position % (kCheckpoint1Interval * kCheckpoint2Interval) == 0) {
-                    C2.emplace_back(std::move(entry_y_bits));
-                }
-                deltas_to_write.clear();
-                ++num_C1_entries;
-            } else {
-                if (entry_y == prev_y) {
-                    deltas_to_write.push_back(0);
-                } else {
-                    deltas_to_write.push_back(entry_y - prev_y);
-                }
-                prev_y = entry_y;
-            }
-        }
-        Encoding::ANSFree(kC3R);
-        delete res.table7_sm;
-
-        // Writes the final park to disk
-        memset(P7_entry_buf, 0, P7_park_size);
-        to_write_p7.ToBytes(P7_entry_buf);
-
-        tmp2_disk.Write(final_file_writer_3, (P7_entry_buf), P7_park_size);
-        final_file_writer_3 += P7_park_size;
-
-        if (!deltas_to_write.empty()) {
-            size_t num_bytes = Encoding::ANSEncodeDeltas(deltas_to_write, kC3R, C3_entry_buf + 2);
-            memset(C3_entry_buf + num_bytes + 2, 0, size_C3 - (num_bytes + 2));
-            final_file_writer_2 = begin_byte_C3 + (num_C1_entries - 1) * size_C3;
-
-            // Write the size
-            Util::IntToTwoBytes(C3_entry_buf, num_bytes);
-
-            tmp2_disk.Write(final_file_writer_2, (C3_entry_buf), size_C3);
-            final_file_writer_2 += size_C3;
-            Encoding::ANSFree(kC3R);
-        }
-
-        Bits(0, Util::ByteAlign(k)).ToBytes(C1_entry_buf);
-        tmp2_disk.Write(final_file_writer_1, (C1_entry_buf), Util::ByteAlign(k) / 8);
-        final_file_writer_1 += Util::ByteAlign(k) / 8;
-        std::cout << "\tFinished writing C1 and C3 tables" << std::endl;
-        std::cout << "\tWriting C2 table" << std::endl;
-
-        for (Bits &C2_entry : C2) {
-            C2_entry.ToBytes(C1_entry_buf);
-            tmp2_disk.Write(final_file_writer_1, (C1_entry_buf), Util::ByteAlign(k) / 8);
-            final_file_writer_1 += Util::ByteAlign(k) / 8;
-        }
-        Bits(0, Util::ByteAlign(k)).ToBytes(C1_entry_buf);
-        tmp2_disk.Write(final_file_writer_1, (C1_entry_buf), Util::ByteAlign(k) / 8);
-        final_file_writer_1 += Util::ByteAlign(k) / 8;
-        std::cout << "\tFinished writing C2 table" << std::endl;
-
-        delete[] C3_entry_buf;
-        delete[] C1_entry_buf;
-        delete[] P7_entry_buf;
-
-        final_file_writer_1 = res.header_size - 8 * 3;
-        uint8_t table_pointer_bytes[8];
-
-        // Writes the pointers to the start of the tables, for proving
-        for (int i = 8; i <= 10; i++) {
-            Util::IntToEightBytes(table_pointer_bytes, res.final_table_begin_pointers[i]);
-            tmp2_disk.Write(final_file_writer_1, table_pointer_bytes, 8);
-            final_file_writer_1 += 8;
-        }
-
-        std::cout << "\tFinal table pointers:" << std::endl << std::hex;
-
-        for (int i = 1; i <= 10; i++) {
-            std::cout << "\t" << (i < 8 ? "P" : "C") << (i < 8 ? i : i - 7);
-            std::cout << ": 0x" << res.final_table_begin_pointers[i] << std::endl;
-        }
-        std::cout << std::dec;
     }
 };
