+++ src/pos_constants.hpp
@@ -20,6 +20,19 @@
 // Unique plot id which will be used as a ChaCha8 key, and determines the PoSpace.
 const uint32_t kIdLen = 32;
 
+// Distance between matching entries is stored in the offset
+const uint32_t kOffsetSize = 10;
+
+// Max matches a single entry can have, used for hardcoded memory allocation
+const uint32_t kMaxMatchesSingleEntry = 30;
+const uint32_t kMinBuckets = 16;
+const uint32_t kMaxBuckets = 128;
+
+// During backprop and compress, the write pointer is ahead of the read pointer
+// Note that the large the offset, the higher these values must be
+const uint32_t kReadMinusWrite = 1U << kOffsetSize;
+const uint32_t kCachedPositionsSize = kReadMinusWrite * 4;
+
 // Must be set high enough to prevent attacks of fast plotting
 const uint32_t kMinPlotSize = 18;
 
@@ -28,6 +41,8 @@ const uint32_t kMaxPlotSize = 50;
 
 // The amount of spare space used for sort on disk (multiplied time memory buffer size)
 const uint32_t kSpareMultiplier = 5;
+
+const double kMemSortProportion = 0.75;
 
 // How many f7s per C1 entry, and how many C1 entries per C2 entry
 const uint32_t kCheckpoint1Interval = 10000;
